fixed some compiler warnings

+++ VPN/Evan/DataTypes/LandmarkSet.cpp
@@ -7,7 +7,7 @@ bool LandmarkSet::setFromContainer(DataTypePtr& specimens, unsigned int partInde
     if(specimens.isValid())
     {
         Specimens* spec = (Specimens*)specimens.getPtr();
-        if(partIndex >= spec->getSize() || partIndex < 0)
+        if(partIndex >= spec->getSize() || ( static_cast<int>(partIndex) < 0 ) )
             return false;
         *this = *(spec->getLandmarkSet(partIndex));
     }
